Salida del decoder en buffers estereo (Left y Right). Comienzo de ecualizador, todavía no anda.

+++ source/LEDDisplay.h
@@ -20,7 +20,7 @@ extern LEDDisplay_ LEDDisplay;
 struct LEDDisplay_{
 	void (*init)();
 
-	void (*setVumeter)(float* input,unsigned len,int mode);
+	void (*setVumeter)(short* input,unsigned len,int mode);
 
 	float* (*getEqualizer)();
 };
